change from uin8 to uiint16 for offchain_derivation_version

+++ canoe/bindings/src/lib.rs
@@ -115,7 +115,7 @@ sol! {
         NonSignerStakesAndSignature nonSignerStakesAndSignature;
         // signed quorum numbers contains all the quorum signed in a batch by the eigenda network
         bytes signedQuorumNumbers;
-        uint8 offchainDerivationVersion;
+        uint16 offchainDerivationVersion;
     }
 
     // see contract status code
